Use a new jemalloc build tag
